Remove nginx again and make it simple
So the story was that I didn't see gunicorn logs in the console, but I
saw them on render, and I though that it's because they have some
additional logs from reverse proxy or something.

So to make dev and prod environment similar I went and made myself
reverse proxy as well, and disabled console logs for request info, while
still logging them in the logfile.

And then imagine my surprise when I discovered that render.com enables
accesslog for gunicorn for me, and that it really was gunicorn, not some
strange reverse proxy and if I disable console logs for gunicorn, I
wouldn't see any request logs.

Yeah, wild stuff.

Anyway, in the end, I enabled them back and yeah, removed nginx.

Gosh.