Implement SignUpView as a FormView subclass
Well, it was easier than I thought, because I already follow established
patterns.

I can't say I like what I've done, but I can't say I dislike it either.
This code is easy to read and write, BUT, you need to know what all of
these mixins are doing and what's happening under the hood.

Like, with this commit, we no longer send 400 on incorrect forms,
because default form_invalid() method simply returns the form back with
status code 200.
But maybe that's the way things should be? Idk.

It's complicated and I lack experience, but it works, it passes
automated tests and I tried it once myself.