Reference a user in comment model
Previously it stored a username as a string, but yeah, using foreign key
is much better, I think.
It even has smart on_delete behaviour, sort of, where it erases the user
info, but keep the comment intact.

Not that I think it would be used that often, as Django recommends
moving users into inactive state instead.